float not decimal

+++ db/migrate/20170724025031_add_discount_to_discounts.rb
@@ -1,5 +1,5 @@
 class AddDiscountToDiscounts < ActiveRecord::Migration
   def change
-    add_column :discounts, :discount, :decimal
+    add_column :discounts, :discount, :float
   end
 end
